EDSC-3148: PR Feedback
