+++ static/src/js/components/OrderStatus/OrderStatusItem.js
@@ -12,367 +12,32 @@ import { getApplicationConfig } from '../../../../../sharedUtils/config'
 import { getStateFromOrderStatus, aggregatedOrderStatus, formatOrderStatus } from '../../../../../sharedUtils/orderStatus'
 import { pluralize } from '../../util/pluralize'
 import { commafy } from '../../util/commafy'
-import { generateDownloadScript } from '../../util/files/generateDownloadScript'
-import { getFilenameFromPath } from '../../util/getFilenameFromPath'
 
 import PortalLinkContainer from '../../containers/PortalLinkContainer/PortalLinkContainer'
 import Button from '../Button/Button'
 import EDSCTabs from '../EDSCTabs/EDSCTabs'
 import OrderProgressList from '../OrderProgressList/OrderProgressList'
-import TextWindowActions from '../TextWindowActions/TextWindowActions'
-import CopyableText from '../CopyableText/CopyableText'
+import STACJsonPanel from './OrderStatusItem/STACJsonPanel'
+import DownloadFilesPanel from './OrderStatusItem/DownloadFilesPanel'
+import S3LinksPanel from './OrderStatusItem/S3LinksPanel'
+import DownloadScriptPanel from './OrderStatusItem/DownloadScriptPanel'
 
 import ProgressRing from '../ProgressRing/ProgressRing'
 
 import './OrderStatusItem.scss'
 
-export const STACJsonPanel = ({
-  accessMethodType,
-  stacLinks,
-  retrievalId,
-  granuleCount,
-  stacLinksIsLoading
-}) => {
-  const downloadFileName = `${retrievalId}-${accessMethodType}-STAC.txt`
-
-  return stacLinks.length > 0 ? (
-    <>
-      <div className="order-status-item__tab-intro">
-        <span className="order-status-item__status-text">
-          {
-            stacLinksIsLoading
-              ? `Retrieving STAC links for ${commafy(granuleCount)} ${pluralize('granule', granuleCount)}...`
-              : `Retrieved ${stacLinks.length} STAC ${pluralize('links', stacLinks.length)} for ${commafy(granuleCount)} ${pluralize('granule', granuleCount)}`
-          }
-        </span>
-      </div>
-      <TextWindowActions
-        id={`links-${retrievalId}`}
-        fileContents={stacLinks.join('\n')}
-        fileName={downloadFileName}
-        clipboardContents={stacLinks.join('\n')}
-        modalTitle="STAC Links"
-      >
-        <ul className="download-links-panel__list">
-          {
-            stacLinks.map((link, i) => {
-              const key = `link_${i}`
-              return (
-                <li key={key}>
-                  <a href={link}>{link}</a>
-                </li>
-              )
-            })
-        }
-        </ul>
-      </TextWindowActions>
-    </>
-  )
-    : (
-      <div className="order-status-item__tab-intro">
-        STAC links will become available once the order has finished processing
-      </div>
-    )
-}
-
-STACJsonPanel.propTypes = {
-  accessMethodType: PropTypes.string.isRequired,
-  stacLinks: PropTypes.arrayOf(
-    PropTypes.string
-  ).isRequired,
-  retrievalId: PropTypes.string.isRequired,
-  granuleCount: PropTypes.number.isRequired,
-  stacLinksIsLoading: PropTypes.bool.isRequired
-}
-
-export const DownloadFilesPanel = ({
-  accessMethodType,
-  downloadLinks,
-  retrievalId,
-  granuleCount,
-  granuleLinksIsLoading,
-  showTextWindowActions
-}) => {
-  const downloadFileName = `${retrievalId}-${accessMethodType}.txt`
-
-  return downloadLinks.length > 0 ? (
-    <>
-      <div className="order-status-item__tab-intro">
-        <span className="order-status-item__status-text">
-          {
-            granuleLinksIsLoading
-              ? `Retrieving files for ${commafy(granuleCount)} ${pluralize('granule', granuleCount)}...`
-              : `Retrieved ${downloadLinks.length} ${pluralize('file', downloadLinks.length)} for ${commafy(granuleCount)} ${pluralize('granule', granuleCount)}`
-          }
-        </span>
-      </div>
-      <TextWindowActions
-        id={`links-${retrievalId}`}
-        fileContents={downloadLinks.join('\n')}
-        fileName={downloadFileName}
-        clipboardContents={downloadLinks.join('\n')}
-        modalTitle="Download Files"
-        disableCopy={!showTextWindowActions}
-        disableSave={!showTextWindowActions}
-      >
-        <ul className="download-links-panel__list">
-          {
-            downloadLinks.map((link, i) => {
-              const key = `link_${i}`
-              return (
-                <li key={key}>
-                  <a href={link}>{link}</a>
-                </li>
-              )
-            })
-          }
-        </ul>
-      </TextWindowActions>
-    </>
-  )
-    : (
-      <div className="order-status-item__tab-intro">
-        The download files will become available once the order has finished processing
-      </div>
-    )
-}
-
-DownloadFilesPanel.defaultProps = {
-  showTextWindowActions: true
-}
-
-DownloadFilesPanel.propTypes = {
-  accessMethodType: PropTypes.string.isRequired,
-  downloadLinks: PropTypes.arrayOf(
-    PropTypes.string
-  ).isRequired,
-  retrievalId: PropTypes.string.isRequired,
-  granuleCount: PropTypes.number.isRequired,
-  granuleLinksIsLoading: PropTypes.bool.isRequired,
-  showTextWindowActions: PropTypes.bool
-}
-
-export const S3LinksPanel = ({
-  accessMethodType,
-  directDistributionInformation,
-  s3Links,
-  retrievalId,
-  granuleCount,
-  granuleLinksIsLoading,
-  showTextWindowActions
-}) => {
-  const downloadFileName = `${retrievalId}-${accessMethodType}-s3.txt`
-
-  const {
-    region,
-    s3BucketAndObjectPrefixNames = [],
-    s3CredentialsApiDocumentationUrl,
-    s3CredentialsApiEndpoint
-  } = directDistributionInformation
-
-  return s3Links.length > 0 ? (
-    <>
-      <div className="order-status-item__tab-intro">
-        {
-          region && (
-            <>
-              <p>{`Direct cloud access for this collection is available in the ${region} region in AWS S3.`}</p>
-              <div className="order-status-item__direct-distribution-information">
-                <div className="order-status-item__direct-distribution-item">
-                  {'Region'}
-                  <CopyableText
-                    className="order-status-item__direct-distribution-item-value"
-                    text={region}
-                    label="Copy to clipboard"
-                    successMessage="Copied the AWS S3 region"
-                    failureMessage="Could not copy the AWS S3 region"
-                  />
-                </div>
-                <div className="order-status-item__direct-distribution-item">
-                  {'Bucket/Object Prefix'}
-                  {s3BucketAndObjectPrefixNames.map((bucketAndObjPrefix, i) => (
-                    <React.Fragment key={`${region}_${bucketAndObjPrefix}`}>
-                      <CopyableText
-                        className="order-status-item__direct-distribution-item-value"
-                        text={bucketAndObjPrefix}
-                        label="Copy to clipboard"
-                        successMessage="Copied the AWS S3 Bucket/Object Prefix"
-                        failureMessage="Could not copy the AWS S3 Bucket/Object Prefix"
-                      />
-                      {i !== s3BucketAndObjectPrefixNames.length - 1 && ', '}
-                    </React.Fragment>
-                  ))}
-                </div>
-                <div className="order-status-item__direct-distribution-item">
-                  {'AWS S3 Credentials'}
-                  <span className="order-status-item__direct-distribution-item-value">
-                    <a
-                      className="link link--external"
-                      href={s3CredentialsApiEndpoint}
-                      rel="noopener noreferrer"
-                      target="_blank"
-                    >
-                      Get AWS S3 Credentials
-                    </a>
-                    <a
-                      className="link link--separated link--external"
-                      href={s3CredentialsApiDocumentationUrl}
-                      rel="noopener noreferrer"
-                      target="_blank"
-                    >
-                      View Documentation
-                    </a>
-                  </span>
-                </div>
-              </div>
-              <span className="order-status-item__status-text">
-                {
-                  granuleLinksIsLoading
-                    ? `Retrieving objects for ${commafy(granuleCount)} ${pluralize('granule', granuleCount)}...`
-                    : `Retrieved ${s3Links.length} ${pluralize('object', s3Links.length)} for ${commafy(granuleCount)} ${pluralize('granule', granuleCount)}`
-                }
-              </span>
-            </>
-          )
-        }
-      </div>
-      <TextWindowActions
-        id={`links-${retrievalId}`}
-        fileContents={s3Links.join('\n')}
-        fileName={downloadFileName}
-        clipboardContents={s3Links.join('\n')}
-        modalTitle="AWS S3 Access"
-        disableCopy={!showTextWindowActions}
-        disableSave={!showTextWindowActions}
-      >
-        <ul className="download-links-panel__list">
-          {
-            s3Links.map((path, i) => {
-              const key = `link_${i}`
-              const s3LinkTitle = getFilenameFromPath(path)
-
-              return (
-                <li key={key}>
-                  <CopyableText
-                    text={path}
-                    label="Copy AWS S3 path to clipboard"
-                    successMessage={() => `Copied AWS S3 path for: ${s3LinkTitle}`}
-                    failureMessage={() => `Could not copy AWS S3 path for: ${s3LinkTitle}`}
-                  />
-                </li>
-              )
-            })
-          }
-        </ul>
-      </TextWindowActions>
-    </>
-  )
-    : (
-      <div className="order-status-item__tab-intro">
-        The AWS S3 objects will become available once the order has finished processing
-      </div>
-    )
-}
-
-S3LinksPanel.defaultProps = {
-  showTextWindowActions: true
-}
-
-S3LinksPanel.propTypes = {
-  accessMethodType: PropTypes.string.isRequired,
-  directDistributionInformation: PropTypes.shape({}).isRequired,
-  s3Links: PropTypes.arrayOf(
-    PropTypes.string
-  ).isRequired,
-  retrievalId: PropTypes.string.isRequired,
-  granuleCount: PropTypes.number.isRequired,
-  granuleLinksIsLoading: PropTypes.bool.isRequired,
-  showTextWindowActions: PropTypes.bool
-}
-
-
-export const DownloadScriptPanel = ({
-  accessMethodType,
-  earthdataEnvironment,
-  downloadLinks,
-  retrievalCollection,
-  retrievalId,
-  granuleCount,
-  granuleLinksIsLoading
-}) => {
-  const downloadFileName = `${retrievalId}-${accessMethodType}.sh`
-
-  return downloadLinks.length > 0
-    ? (
-      <>
-        <div className="order-status-item__tab-intro">
-          <p className="collection-download-display__intro mt-2">
-            <strong>Linux: </strong>
-            { 'You must first make the script an executable by running the line \'chmod 777 download.sh\' from the command line. After that is complete, the file can be executed by typing \'./download.sh\'. ' }
-            { 'For a detailed walk through of this process, please reference this ' }
-            <a href="https://wiki.earthdata.nasa.gov/display/EDSC/How+To%3A+Use+the+Download+Access+Script">How To guide</a>
-            { '.' }
-          </p>
-          <p>
-            <strong>Windows: </strong>
-            {
-              'The file can be executed within Windows by first installing a Unix-like command line utility such as '
-            }
-            <a href="https://www.cygwin.com/">Cygwin</a>
-            {
-              '. After installing Cygwin (or a similar utility), run the line \'chmod 777 download.sh\' from the utility\'s command line, and then execute by typing \'./download.sh\'.'
-            }
-          </p>
-          <span className="order-status-item__status-text">
-            {
-              granuleLinksIsLoading
-                ? `Retrieving files for ${commafy(granuleCount)} ${pluralize('granule', granuleCount)}...`
-                : `Retrieved ${downloadLinks.length} files for ${commafy(granuleCount)} ${pluralize('granule', granuleCount)}`
-            }
-          </span>
-        </div>
-        <TextWindowActions
-          id={`script-${retrievalId}`}
-          fileContents={generateDownloadScript(
-            downloadLinks,
-            retrievalCollection,
-            earthdataEnvironment
-          )}
-          fileName={downloadFileName}
-          clipboardContents={generateDownloadScript(
-            downloadLinks,
-            retrievalCollection,
-            earthdataEnvironment
-          )}
-          modalTitle="Download Script"
-        >
-          <pre className="download-links-panel__pre">
-            {
-              generateDownloadScript(downloadLinks, retrievalCollection, earthdataEnvironment)
-            }
-          </pre>
-        </TextWindowActions>
-      </>
-    )
-    : (
-      <div className="order-status-item__tab-intro">
-          The download script will become available once the order has finished processing
-      </div>
-    )
-}
-
-DownloadScriptPanel.propTypes = {
-  accessMethodType: PropTypes.string.isRequired,
-  earthdataEnvironment: PropTypes.string.isRequired,
-  downloadLinks: PropTypes.arrayOf(
-    PropTypes.string
-  ).isRequired,
-  retrievalCollection: PropTypes.shape({}).isRequired,
-  retrievalId: PropTypes.string.isRequired,
-  granuleCount: PropTypes.number.isRequired,
-  granuleLinksIsLoading: PropTypes.bool.isRequired
-}
-
+/**
+ * Renders OrderStatusItem.
+ * @param {Object} arg0 - The props passed into the component.
+ * @param {Boolean} arg0.defaultOpen - Sets the item open on intital render.
+ * @param {String} arg0.earthdataEnvironment - The earthdata environment.
+ * @param {Object} arg0.granuleDownload - granuleDownload state.
+ * @param {Object} arg0.collection - The collection state.
+ * @param {Object} arg0.match - match parameter from React Router.
+ * @param {Function} arg0.onChangePath - Callback function to change the current path.
+ * @param {Function} arg0.onFetchRetrievalCollection - Callback function to fetch a retrieval.
+ * @param {Function} arg0.onFetchRetrievalCollectionGranuleLinks - Callback function to fetch a links for a retrieval collection.
+*/
 export class OrderStatusItem extends PureComponent {
   constructor(props) {
     super(props)
